+++ http/src/test/kotlin/at/asitplus/wallet/backend/IssuingInternalAuthorizationServerTest.kt
@@ -33,7 +33,7 @@ import at.asitplus.wallet.lib.openid.AuthenticationResponseResult
 import at.asitplus.wallet.mdl.MobileDrivingLicenceScheme
 import at.asitplus.wallet.por.PowerOfRepresentationDataElements
 import at.asitplus.wallet.por.PowerOfRepresentationScheme
-import at.asitplus.wallet.taxid.TaxId2025Scheme
+import at.asitplus.wallet.taxid.TaxIdScheme
 import com.benasher44.uuid.uuid4
 import io.kotest.matchers.collections.shouldContain
 import io.kotest.matchers.ints.shouldBeGreaterThan
@@ -212,8 +212,8 @@ class IssuingInternalAuthorizationServerTest {
     }
 
     @Test
-    fun taxid2025_ok() = runTest {
-        val requestOptions = WalletService.RequestOptions(TaxId2025Scheme, SD_JWT)
+    fun taxid_ok() = runTest {
+        val requestOptions = WalletService.RequestOptions(TaxIdScheme, SD_JWT)
 
         val credential = loadCredential(requestOptions)